Fix header pulled out of page flow
Header css position absolute was added so it wont add to the
height of the page. This was needed because the app did not fit in a
single page. It was caused by the main tag height is set to 100%
view height so it can be centered in the page.